+++ Source/SwiftLintBuiltInRules/Rules/Idiomatic/FileNameRule.swift
@@ -1,3 +1,4 @@
+import SwiftLintCore
 import SwiftSyntax
 
 struct FileNameRule: OptInRule, SourceKitFreeRule {
@@ -12,6 +13,7 @@ struct FileNameRule: OptInRule, SourceKitFreeRule {
 
     func validate(file: SwiftLintFile) -> [StyleViolation] {
         guard let filePath = file.path,
+            !shouldExclude(filePath: filePath),
             case let fileName = filePath.bridge().lastPathComponent,
             !configuration.excluded.contains(fileName) else {
             return []
@@ -54,6 +56,16 @@ struct FileNameRule: OptInRule, SourceKitFreeRule {
                 location: Location(file: filePath, line: 1)
             ),
         ]
+    }
+}
+
+// MARK: - For `excluded` option
+private extension FileNameRule {
+    func shouldExclude(filePath: String) -> Bool {
+        configuration.excluded.compactMap {
+            let regex = try? RegularExpression(pattern: "^\($0)$")
+            return regex?.regex.firstMatch(in: filePath, options: [], range: filePath.fullNSRange)
+        }.isNotEmpty
     }
 }