+++ CHANGELOG.md
@@ -14,7 +14,9 @@
 
 ### Enhancements
 
-* None.
+* Add support for regex pattern in the `excluded` option of the `file_name` rule.  
+  [Ueeek](https://github.com/Ueeek)
+  [#6066](https://github.com/realm/SwiftLint/issues/6066)
 
 ### Bug Fixes